requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine learning utilities
scikit-learn>=1.3.0

# Hyperparameter search
optuna>=3.5.0

# Plotting & visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & logging
pyyaml>=6.0
tensorboard>=2.14.0

# Statistical metrics
statsmodels>=0.14.0

# Testing
pytest>=7.4.0

# Progress bars
tqdm>=4.66.0

# Optional graph utilities (if extended later)
networkx>=3.1

# Optional experiment tracking (safe fallback if unused)
wandb>=0.16.0